+++ assets/css/graphics/_icons.scss
@@ -84,3 +84,8 @@
     font-family: fontAwesome;
     color: white;
 }
+
+.avatar {
+    height: 75px;
+    padding: 5px;
+}
