+++ views/desktop/pages/administration.html.twig
@@ -127,8 +127,7 @@
                 </div>
                 <div class="box-body text-center">
                     <a class="btn btn-app" href="index.php?v=d&p=general"><i class="fas fa-wrench"></i>{{ 'page-title.general'|trans }} </a>
-                    <a class="btn btn-app" href="index.php?v=d&p=custom"><i class="fas fa-pencil-square"></i>{{ 'page-title.customization'|trans }}</a>
-                    <a class="btn btn-app" href="index.php?v=d&p=profils"><i class="fas fa-briefcase"></i>{{ 'page-title.profil'|trans }}</a>
+                    <a class="btn btn-app" href="index.php?v=d&p=profils"><i class="fas fa-users-cog"></i>{{ 'page-title.profil'|trans }}</a>
                     <a class="btn btn-app" href="index.php?v=d&p=commandes"><i class="fas fa-hand-o-right"></i>{{ 'page-title.commands'|trans }}</a>
                     <a class="btn btn-app" href="index.php?v=d&p=links"><i class="fas fa-sitemap"></i>{{ 'page-title.links'|trans }}</a>
                     <a class="btn btn-app" href="index.php?v=d&p=interact_config"><i class="fas fa-comments"></i>{{ 'page-title.interacts'|trans }}</a>
